# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SaaS Multi-Niche API",
        "description": "API RESTful pour système de réservation multi-niche (paragliding, surfing, diving, etc.)",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server (Développement)"
        },
        {
            "url": "https://api.example.com",
            "description": "API Server (Production)"
        }
    ],
    "paths": {
...
        "/api/v1/admin/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Liste des réservations (Admin)",
                "description": "Retourne la liste des réservations avec filtres et pagination",
                "operationId": "adminListReservations",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrer par statut",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "activity_type",
                        "in": "query",
                        "description": "Filtrer par type d'activité",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "instructor_id",
                        "in": "query",
                        "description": "Filtrer par instructeur",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Date de début",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Date de fin",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche (email, nom, UUID)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des réservations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "429": {
                        "description": "Rate limit atteint"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {
                        "organization": []
                    }
                ]
            }
        },
        "/api/v1/admin/reservations/{id}": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Détails d'une réservation (Admin)",
                "description": "Retourne les détails complets d'une réservation",
                "operationId": "adminGetReservation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la réservation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Reservation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Réservation non trouvée"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "429": {
                        "description": "Rate limit atteint"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {
                        "organization": []
                    }
                ]
            }
        },
        "/api/v1/admin/reservations/{id}/assign": {
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Assigner date et ressources à une réservation",
                "description": "Planifie une réservation en assignant un instructeur, un site et des équipements",
                "operationId": "assignReservation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "scheduled_at",
                                    "instructor_id"
                                ],
                                "properties": {
                                    "scheduled_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-12-25T10:00:00Z"
                                    },
                                    "instructor_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "site_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "equipment_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "vehicle_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réservation assignée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Reservation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur lors de l'assignation"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "429": {
                        "description": "Rate limit atteint"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {
                        "organization": []
                    }
                ]
            }
        },
...
        "/api/v1/payments/capture": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Capturer un paiement autorisé",
                "description": "Capture un paiement qui a été préalablement autorisé (empreinte)",
                "operationId": "capturePayment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "payment_id"
                                ],
                                "properties": {
                                    "payment_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "amount": {
                                        "description": "Montant à capturer (optionnel, capture totale si omis)",
                                        "type": "number",
                                        "format": "float",
                                        "example": 120,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paiement capturé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Payment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "429": {
                        "description": "Rate limit atteint"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {
                        "organization": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum Bearer Token. Obtenez le token via /api/v1/auth/login",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "organization": {
                "type": "apiKey",
                "description": "ID de l'organisation (tenant). Requis pour l'isolation multi-tenant.",
                "name": "X-Organization-ID",
                "in": "header"
            }
        }
    },
...
}